Fixing Bugs

+++ Uber Eats Database Project/ORDER.cs
@@ -20,7 +20,7 @@ namespace Uber_Eats_Database_Project
             this.ORDER_FOOD = new HashSet<ORDER_FOOD>();
         }
     
-        public decimal ORDER_ID { get; set; }
+        public int ORDER_ID { get; set; }
         public string TIP { get; set; }
         public Nullable<System.DateTime> ORDER_DATE { get; set; }
         public string PAYMENT_TYPE { get; set; }
